Consolidate all object-storage related models into one module.
