--- marketplace_standard_app_api/models/data_sink.py
@@ -1,10 +1,0 @@
-from datetime import datetime
-from typing import NewType
-
-from pydantic import UUID4, BaseModel
-
-DatasetId = NewType("DatasetId", UUID4)
-
-
-class DatasetCreateResponse(BaseModel):
-    last_modified: datetime
